+++ sdk/python/kfp/components/component_factory.py
@@ -17,7 +17,7 @@ import itertools
 import pathlib
 import re
 import textwrap
-from typing import Callable, List, Optional, Tuple, Type, Union
+from typing import Callable, List, Optional, Tuple
 import warnings
 
 import docstring_parser
@@ -28,7 +28,6 @@ from kfp.components import python_component
 from kfp.components import structures
 from kfp.components.container_component_artifact_channel import \
     ContainerComponentArtifactChannel
-from kfp.components.types import artifact_types
 from kfp.components.types import custom_artifact_types
 from kfp.components.types import type_annotations
 from kfp.components.types import type_utils
@@ -144,17 +143,6 @@ def _maybe_make_unique(name: str, names: List[str]):
     raise RuntimeError('Too many arguments with the name {}'.format(name))
 
 
-def is_list_of_artifacts(
-    type_var: Union[Type[List[artifact_types.Artifact]],
-                    Type[artifact_types.Artifact]]
-) -> bool:
-    # the type annotation for this function's `type_var` parameter may not actually be a subclass of the KFP SDK's Artifact class for custom artifact types
-    return hasattr(
-        type_var, '__origin__'
-    ) and type_var.__origin__ == list and type_annotations.is_artifact_class(
-        type_var.__args__[0])
-
-
 def extract_component_interface(
         func: Callable,
         containerized: bool = False) -> structures.ComponentSpec:
@@ -184,7 +172,8 @@ def extract_component_interface(
 
             # parameter_type is a type like typing_extensions.Annotated[kfp.components.types.artifact_types.Artifact, <class 'kfp.components.types.type_annotations.OutputAnnotation'>] OR typing_extensions.Annotated[typing.List[kfp.components.types.artifact_types.Artifact], <class 'kfp.components.types.type_annotations.OutputAnnotation'>]
 
-            is_artifact_list = is_list_of_artifacts(parameter_type.__origin__)
+            is_artifact_list = type_annotations.is_list_of_artifacts(
+                parameter_type.__origin__)
 
             parameter_type = type_annotations.get_io_artifact_class(
                 parameter_type)
@@ -266,7 +255,7 @@ def extract_component_interface(
                 output_name = _maybe_make_unique(field_name, output_names)
                 output_names.add(output_name)
                 type_var = field_annotations.get(field_name)
-                if is_list_of_artifacts(type_var):
+                if type_annotations.is_list_of_artifacts(type_var):
                     raise ValueError(
                         f'Cannot use output lists of artifacts in NamedTuple return annotations. Got output list of artifacts annotation for NamedTuple field `{field_name}`.'
                     )
@@ -299,7 +288,7 @@ def extract_component_interface(
             #   `def func(output_path: OutputPath()) -> str: ...`
             output_names.add(output_name)
             return_ann = signature.return_annotation
-            if is_list_of_artifacts(return_ann):
+            if type_annotations.is_list_of_artifacts(return_ann):
                 artifact_cls = return_ann.__args__[0]
                 output_spec = structures.OutputSpec(
                     type=type_utils.create_bundled_artifact_type(
